Add support for trackers on maintenance page
