+++ context/app/static/js/maintenance/index.html
@@ -4,6 +4,10 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <title>Maintenance | HuBMAP</title>
+    <script>
+      const sentryDsn = '{{ sentry_dsn }}';    
+      const sentryEnv = '{{ sentry_env }}';
+    </script>
   </head>
   <body>
     <div id="react-content"></div>
